feat: 비즈니스 로직 및 정리
- 객체 모델 변경에 따라서 로직이 수정될 가능성이 있습니다

+++ peauty-payment/src/main/java/com/peauty/payment/presentation/dto/OrderRequest.java
@@ -3,7 +3,7 @@ package com.peauty.payment.presentation.dto;
 import com.peauty.payment.business.dto.OrderCommand;
 
 public record OrderRequest(
-        Integer cost
+        Long cost
 ) {
     public OrderCommand toCommand() {
         return OrderCommand.builder()
